feat: Add HTTP cache provider
We'd like to use plain HTTP for remote object storage without
S3 specific calculations.
