+++ src/cache/remote_cache.cpp
@@ -20,6 +20,7 @@
 #include <cache/remote_cache.hpp>
 
 #include <base/debug_utils.hpp>
+#include <cache/http_cache_provider.hpp>
 #include <cache/redis_cache_provider.hpp>
 #include <config/configuration.hpp>
 
@@ -71,7 +72,9 @@ bool remote_cache_t::connect() {
 
   // Select an apropriate cache provider.
   m_provider = nullptr;
-  if (protocol == "redis") {
+  if (protocol == "http") {
+    m_provider = new http_cache_provider_t();
+  } else if (protocol == "redis") {
     m_provider = new redis_cache_provider_t();
 #ifdef ENABLE_S3
   } else if (protocol == "s3") {
